reafactor: types
