+++ app/project-manager-shim/src/index.ts
@@ -1,3 +1,2 @@
 export * from './filesystem.js'
-export * from './types.js'
 export * from './projectManagement.js'
