--- app/project-manager-shim/src/types.ts
@@ -1,61 +1,0 @@
-/** Details of a project. */
-export interface ProjectMetadata {
-  /** The name of the project. */
-  readonly name: string
-  /** The namespace of the project. */
-  readonly namespace: string
-  /** The project id. */
-  readonly id: string
-  /**
-   * The Enso Engine version to use for the project, represented by a semver version
-   * string.
-   *
-   * If the edition associated with the project could not be resolved, the
-   * engine version may be missing.
-   */
-  readonly engineVersion?: string
-  /** The project creation time. */
-  readonly created: string
-  /** The last opened datetime. */
-  readonly lastOpened?: string
-}
-
-/** Attributes of a file or folder. */
-export interface Attributes {
-  readonly creationTime: string
-  readonly lastAccessTime: string
-  readonly lastModifiedTime: string
-  readonly byteSize: number
-}
-
-/** Metadata for an arbitrary file system entry. */
-export type FileSystemEntry = DirectoryEntry | FileEntry | ProjectEntry
-
-/** The discriminator value for {@link FileSystemEntry}. */
-export enum FileSystemEntryType {
-  DirectoryEntry = 'DirectoryEntry',
-  ProjectEntry = 'ProjectEntry',
-  FileEntry = 'FileEntry',
-}
-
-/** Metadata for a file. */
-export interface FileEntry {
-  readonly type: FileSystemEntryType.FileEntry
-  readonly path: string
-  readonly attributes: Attributes
-}
-
-/** Metadata for a directory. */
-export interface DirectoryEntry {
-  readonly type: FileSystemEntryType.DirectoryEntry
-  readonly path: string
-  readonly attributes: Attributes
-}
-
-/** Metadata for a project. */
-export interface ProjectEntry {
-  readonly type: FileSystemEntryType.ProjectEntry
-  readonly path: string
-  readonly metadata: ProjectMetadata
-  readonly attributes: Attributes
-}
